Small refactoring and interrupt access improvement

+++ lib/blocks/data_types.h
@@ -59,17 +59,17 @@ union AbsPosition {
    * @brief Structure to split position into angle and rotations.
    */
   struct {
-    int32_t angle;      // Angle component of the position.
-    int32_t rotations;  // Rotations component of the position.
+    uint16_t angle;      // Angle component of the position.
+    int16_t rotations;  // Rotations component of the position.
   } split;
-  int64_t position;  // Combined 64-bit position value.
+  int32_t position;  // Combined 64-bit position value.
 
   /**
    * @brief Constructor to initialize the union.
    * @param r Rotations value.
    * @param a Angle value.
    */
-  AbsPosition(int32_t r, int32_t a) {
+  AbsPosition(int16_t r, int16_t a) {
     split.rotations = r;
     split.angle = a;
   }
